% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntms.R
\name{export_neatms}
\alias{export_neatms}
\title{Export aligned feature table from patRoon to NeatMS
This function exports the aligned feature table from patRoon to a format compatible with NeatMS. It extracts the necessary information from the XCMSnExp object and the patRoon feature groups, and then combines them into a single dataframe that can be exported as a CSV file for use in NeatMS.}
\usage{
export_neatms(fGroups = fGroupsXCMS, csv_path = "aligned_feature_table.csv")
}
\arguments{
\item{fGroups}{The feature groups object from patRoon (default is fGroupsXCMS).}

\item{csv_path}{The file path where the aligned feature table CSV will be saved (default is "aligned_feature_table.csv").}
}
\value{
A dataframe containing the aligned feature information, which is also saved as a CSV file for NeatMS.
}
\description{
Export aligned feature table from patRoon to NeatMS
This function exports the aligned feature table from patRoon to a format compatible with NeatMS. It extracts the necessary information from the XCMSnExp object and the patRoon feature groups, and then combines them into a single dataframe that can be exported as a CSV file for use in NeatMS.
}
