% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfascreen.R
\name{export_pfascreen}
\alias{export_pfascreen}
\title{Export feature list for PFAScreen
This function exports a feature list from patRoon in a format compatible with PFAScreen. It extracts the necessary information from the MS lists and feature groups, and then combines them into a single dataframe that can be exported as a CSV file for use in PFAScreen.}
\usage{
export_pfascreen(mslists = mslists, path = "results")
}
\arguments{
\item{mslists}{The MS lists object from patRoon (default is mslists).}

\item{path}{The file path where the output CSV file will be saved (default is "results").}
}
\value{
A dataframe containing the feature information, which is also saved as a CSV file for PFAScreen.
}
\description{
Export feature list for PFAScreen
This function exports a feature list from patRoon in a format compatible with PFAScreen. It extracts the necessary information from the MS lists and feature groups, and then combines them into a single dataframe that can be exported as a CSV file for use in PFAScreen.
}
