% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2export.R
\name{generateGNPSInfo}
\alias{generateGNPSInfo}
\title{Export MS2 spectra from patRoon to GNPS format
This function exports MS2 spectra from patRoon v3.0 to a format compatible with GNPS. It generates a feature table and an MGF file containing the MS2 spectra, which can be used for molecular networking in GNPS.}
\usage{
generateGNPSInfo(
  fGroups = fGroups,
  mslists = mslists,
  polarity = c("positive", "negative"),
  path
)
}
\arguments{
\item{fGroups}{The feature groups object from patRoon (default is fGroups).}

\item{mslists}{The MS lists object from patRoon (default is mslists).}

\item{polarity}{The polarity of the MS data, either "positive" or "negative" (default is c("positive", "negative")).}

\item{path}{The file path where the output files will be saved (default is the current working directory).}
}
\value{
A list containing the feature table and the MGF file path for GNPS.
}
\description{
Export MS2 spectra from patRoon to GNPS format
This function exports MS2 spectra from patRoon v3.0 to a format compatible with GNPS. It generates a feature table and an MGF file containing the MS2 spectra, which can be used for molecular networking in GNPS.
}
