% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2export.R
\name{generateSIRIUSmgf}
\alias{generateSIRIUSmgf}
\title{Generate MGF file for SIRIUS from patRoon data
This function generates an MGF file for SIRIUS from patRoon data. It creates a feature table and an MGF file containing the MS2 spectra, which can be used for compound identification in SIRIUS.}
\usage{
generateSIRIUSmgf(
  fGroups = fGroups,
  mslists = mslists,
  polarity = c("positive", "negative"),
  path
)
}
\arguments{
\item{fGroups}{The feature groups object from patRoon (default is fGroups).}

\item{mslists}{The MS lists object from patRoon (default is mslists).}

\item{polarity}{The polarity of the MS data, either "positive" or "negative" (default is c("positive", "negative")).}

\item{path}{The file path where the output MGF file will be saved (default is the current working directory).}
}
\value{
The file path to the generated MGF file for SIRIUS.
}
\description{
Generate MGF file for SIRIUS from patRoon data
This function generates an MGF file for SIRIUS from patRoon data. It creates a feature table and an MGF file containing the MS2 spectra, which can be used for compound identification in SIRIUS.
}
