% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntms.R
\name{import_neatms}
\alias{import_neatms}
\title{Import NeatMS results and create YAML file for removal of noisy features
This function imports the results from NeatMS, combines them with the metadata from patRoon, and creates a YAML file for the removal of noisy features based on the quality labels assigned by NeatMS.}
\usage{
import_neatms(
  ntms_results = "ntms_export.csv",
  anaInfo = anaInfo,
  feature_dataframe = feature_dataframe,
  yaml_path = "model_session.yml"
)
}
\arguments{
\item{ntms_results}{The file path to the NeatMS results CSV file (default is "ntms_export.csv").}

\item{anaInfo}{The analysis information dataframe from patRoon.}

\item{feature_dataframe}{The dataframe containing feature information, including feature IDs and sample names.}

\item{yaml_path}{The file path where the YAML file for removal of noisy features will be saved (default is "model_session.yml").}
}
\value{
A dataframe summarizing feature quality by group, which is also saved as a CSV file for further analysis.
}
\description{
Import NeatMS results and create YAML file for removal of noisy features
This function imports the results from NeatMS, combines them with the metadata from patRoon, and creates a YAML file for the removal of noisy features based on the quality labels assigned by NeatMS.
}
