% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfascreen.R
\name{import_pfascreen}
\alias{import_pfascreen}
\title{Import PFAScreen results and combine with patRoon metadata
This function imports the results from PFAScreen, combines them with the metadata from patRoon, and creates a dataframe that can be used for further analysis. It reads the PFAScreen results from a CSV file, merges them with the feature groups from patRoon, and filters the results based on the number of unique homologues.}
\usage{
import_pfascreen(folder = "results/PFAScreen-main")
}
\arguments{
\item{folder}{The folder path where the PFAScreen results CSV file is located (default is "results/PFAScreen-main"). The function expects a file named "Results_PFAScreen.*\\.csv" within this folder.}
}
\value{
A dataframe containing the combined results from PFAScreen and patRoon, which can be used for further analysis.
}
\description{
Import PFAScreen results and combine with patRoon metadata
This function imports the results from PFAScreen, combines them with the metadata from patRoon, and creates a dataframe that can be used for further analysis. It reads the PFAScreen results from a CSV file, merges them with the feature groups from patRoon, and filters the results based on the number of unique homologues.
}
