% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padel.R
\name{padelDesc}
\alias{padelDesc}
\title{Calculate molecular descriptors using PaDEL-Descriptor
This function calculates molecular descriptors for a given dataframe containing SMILES strings using the PaDEL-Descriptor Java application. It takes a dataframe with a "SMILES" column, runs the PaDEL-Descriptor tool for each SMILES string, and returns a dataframe with the calculated descriptors. The function also includes error handling to manage cases where the descriptor calculation fails for certain SMILES strings.}
\usage{
padelDesc(df, jar_path = "descriptor-cli-0.1a-SNAPSHOT-all.jar")
}
\arguments{
\item{df}{A dataframe containing a column named "SMILES" with the SMILES strings for which descriptors are to be calculated.}

\item{jar_path}{The file path to the PaDEL-Descriptor JAR file (default is "descriptor-cli-0.1a-SNAPSHOT-all.jar").}
}
\value{
A dataframe containing the calculated molecular descriptors for each SMILES string, along with the original SMILES string.
}
\description{
Calculate molecular descriptors using PaDEL-Descriptor
This function calculates molecular descriptors for a given dataframe containing SMILES strings using the PaDEL-Descriptor Java application. It takes a dataframe with a "SMILES" column, runs the PaDEL-Descriptor tool for each SMILES string, and returns a dataframe with the calculated descriptors. The function also includes error handling to manage cases where the descriptor calculation fails for certain SMILES strings.
}
